feat(arc-1500): remove TemplateService
